Fix build: use JDK 21, fix Kotlin compile errors (text drawing, Cover(), tool setter clash, experimental Material3 opt-in)

+++ app/src/main/java/com/binlate/suaanh/editor/EditorRenderer.kt
@@ -23,7 +23,7 @@ object EditorRenderer {
 
         for (layer in layers) {
             when (layer) {
-                is Layer.Stroke -> drawStroke(canvas, layer, base.width, base.height, shorter)
+                is Layer.Stroke -> drawStroke(canvas, layer, base.width, base.height, shorter.toFloat())
                 is Layer.Text -> drawText(canvas, layer, base.width, base.height)
                 is Layer.Cover -> drawCover(canvas, base, layer)
             }
